refactor: Move StringUtil into util package
