+++ src/main/kotlin/com/github/ilovegamecoding/intellijcodexp/util/StringUtil.kt
@@ -1,4 +1,4 @@
-package com.github.ilovegamecoding.intellijcodexp
+package com.github.ilovegamecoding.intellijcodexp.util
 
 object StringUtil {
     fun numberToStringWithCommas(number: Long): String {
